requirements.txt
# Core AI
anthropic>=0.40.0

# Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12
jinja2>=3.1.4

# Resume Parsing
PyPDF2>=3.0.1
python-docx>=1.1.2
pdfplumber>=0.11.4

# Job Search
google-search-results>=2.4.2  # SerpAPI client
httpx>=0.27.0
requests>=2.32.0

# Database
sqlalchemy>=2.0.36
aiosqlite>=0.20.0

# Data Validation
pydantic>=2.10.0
pydantic-settings>=2.6.0

# Environment
python-dotenv>=1.0.1

# CLI
click>=8.1.8
rich>=13.9.0

# Utilities
python-slugify>=8.0.4
